Polish SwiftUI package API and docs
