+++ Pinwheel/README.md
@@ -1,3 +1,5 @@
 # Pinwheel
 
-A description of this package.
+Pinwheel is a SwiftUI-first catalog and playground for inspecting app UI in multiple states, presentations, and device sizes.
+
+See the repository root `README.md` for installation, SwiftUI examples, UIKit compatibility, and migration notes.
